Remove Legacy and Legacy platform and instead move their code to DerivationWithLegacyMacros which use their former code as rule
