+++ chimney/src/main/scala-2/io/scalaland/chimney/internal/compiletime/derivation/transformer/DerivationPlatform.scala
@@ -3,7 +3,8 @@ package io.scalaland.chimney.internal.compiletime.derivation.transformer
 import io.scalaland.chimney.internal.compiletime.{DefinitionsPlatform, DerivationErrors, DerivationResult}
 import io.scalaland.chimney.{partial, PartialTransformer, Transformer}
 
-private[derivation] trait DerivationPlatform extends Derivation with Legacy { this: DefinitionsPlatform =>
+private[derivation] trait DerivationPlatform extends Derivation with DerivationWithLegacyMacros {
+  this: DefinitionsPlatform =>
 
   import c.universe.{Transformer as _, *}
 
@@ -39,5 +40,5 @@ private[derivation] trait DerivationPlatform extends Derivation with Legacy { th
     )
   }
 
-  override protected val rulesAvailableForPlatform: Seq[Rule] = Seq()
+  override protected val rulesAvailableForPlatform: Seq[Rule] = Seq(legacyMacrosRule)
 }
